Refactored ListItemView, added previews with different states moved ListItemVIew to Lists folder
