+++ Packages/Style/Sources/Spacing.swift
@@ -1,6 +1,7 @@
 import Foundation
 
 public struct Spacing {
+    public static let extraSmall: CGFloat = 2
     public static let tiny: CGFloat = 4
     public static let small: CGFloat = 8
     public static let medium: CGFloat = 16
